style: 드롭박스 스타일 컴포 분리
