+++ src/components/Book/DropBox/DropBox.tsx
@@ -1,5 +1,11 @@
-import styled from "@emotion/styled";
 import React, { useRef, useState, useEffect } from "react";
+import {
+  ArrowButton,
+  Button,
+  Container,
+  Dropbox,
+  SelectedButton,
+} from "./DropBoxStyles";
 
 const instruments: string[] = [
   "악기",
@@ -36,9 +42,9 @@ const InstrumentDropBox: React.FC<DropBoxProps> = ({ onClick }) => {
       }
     };
 
-    document.addEventListener("mousedown", handleOutsideClick); // 클릭 이벤트 추가
+    document.addEventListener("mousedown", handleOutsideClick);
     return () => {
-      document.removeEventListener("mousedown", handleOutsideClick); // 이벤트 정리
+      document.removeEventListener("mousedown", handleOutsideClick);
     };
   }, []);
 
@@ -90,9 +96,9 @@ const TeamDropBox: React.FC<DropBoxProps> = ({ onClick }) => {
       }
     };
 
-    document.addEventListener("mousedown", handleOutsideClick); // 클릭 이벤트 추가
+    document.addEventListener("mousedown", handleOutsideClick);
     return () => {
-      document.removeEventListener("mousedown", handleOutsideClick); // 이벤트 정리
+      document.removeEventListener("mousedown", handleOutsideClick);
     };
   }, []);
 
@@ -120,71 +126,5 @@ const TeamDropBox: React.FC<DropBoxProps> = ({ onClick }) => {
     </Container>
   );
 };
-const ArrowButton = styled.div`
-  font-size: 10px;
-  color: #000;
-  background-color: transparent;
-  cursor: pointer;
-`;
-const Container = styled.div`
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  gap: 10px;
-  max-width: 200px;
-`;
-interface DropboxProps {
-  isOpened: boolean;
-}
-const Dropbox = styled.div<DropboxProps>`
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  width: 100px;
-  padding-bottom: 10px;
-  max-height: ${({ isOpened }) => (isOpened ? "200px" : "0px")};
-  transition: max-height 0.3s ease-in-out;
-  border-radius: 10px;
-  overflow-y: scroll;
-  overflow-x: hidden;
-  ::-webkit-scrollbar {
-    display: none;
-  }
-  z-index: 1;
-  gap: 5px;
-  top: 30px;
-  background-color: #fff;
 
-  ${({ isOpened }) =>
-    isOpened &&
-    `
-    box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.05);
-  `}
-`;
-const SelectedButton = styled.button`
-  color: rgb(0, 0, 0);
-  min-width: 100px;
-  white-space: nowrap;
-  height: 30px;
-  margin-right: 5px;
-  background-color: #fff;
-  border-radius: 5px;
-  display: flex;
-  align-items: center;
-  gap: 10px;
-`;
-const Button = styled.button`
-  width: 100%;
-  height: 30px;
-  text-align: left;
-  padding: 8px;
-  background-color: #fff;
-  color: black;
-  margin: 5px;
-  margin-bottom: 0px;
-  &:hover {
-    color: #ffaa00;
-    font-weight: 500;
-  }
-`;
 export { InstrumentDropBox, TeamDropBox };
